Move and update solutions, rename projects
