+++ WinUIGallery.UITests/SampleTestTemplate.cs
