+++ WinUIGallery.UITests/SessionManager.cs
