+++ WinUIGallery.UITests/TestBase.cs
