+++ WinUIGallery.UITests/Tests/CheckBox.cs
