+++ WinUIGallery.UITests/Tests/ComboBox.cs
