+++ WinUIGallery.UITests/Tests/MediaPlayerElement.cs
