+++ WinUIGallery.UITests/Tests/PersonPicture.cs
