+++ WinUIGallery.UITests/Tests/ProgressBar.cs
