+++ WinUIGallery.UITests/Tests/SearchResults.cs
