+++ WinUIGallery.UITests/Tests/Slider.cs
