+++ WinUIGallery.UITests/Tests/ToggleSwitch.cs
