+++ WinUIGallery.UITests/WinUIGallery.UITests.csproj
