+++ WinUIGallery.UnitTests/Assets/LockScreenLogo.scale-200.png
