+++ WinUIGallery.UnitTests/Assets/SplashScreen.scale-200.png
