+++ WinUIGallery.UnitTests/Assets/Square150x150Logo.scale-200.png
