+++ WinUIGallery.UnitTests/Assets/StoreLogo.png
