+++ WinUIGallery.UnitTests/Assets/Wide310x150Logo.scale-200.png
