--- WinUIGallery/WinUIGallery.DesktopWap.sln
@@ -1,62 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.32319.34
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A98961E3-CB85-4AFC-990C-FD9F57043E7D}"
-	ProjectSection(SolutionItems) = preProject
-		..\Directory.Build.props = ..\Directory.Build.props
-	EndProjectSection
-EndProject
-Project("{C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5}") = "WinUIGallery.DesktopWap.Package", "WinUIGallery.DesktopWap.Package.wapproj", "{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WinUIGallery", "WinUIGallery.csproj", "{D78F3B56-340D-44F0-9AF1-320D473BA611}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|ARM64 = Debug|ARM64
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|ARM64 = Release|ARM64
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|ARM64.Build.0 = Debug|ARM64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|ARM64.Deploy.0 = Debug|ARM64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|x64.ActiveCfg = Debug|x64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|x64.Build.0 = Debug|x64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|x64.Deploy.0 = Debug|x64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|x86.ActiveCfg = Debug|x86
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|x86.Build.0 = Debug|x86
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Debug|x86.Deploy.0 = Debug|x86
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|ARM64.ActiveCfg = Release|ARM64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|ARM64.Build.0 = Release|ARM64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|ARM64.Deploy.0 = Release|ARM64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|x64.ActiveCfg = Release|x64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|x64.Build.0 = Release|x64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|x64.Deploy.0 = Release|x64
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|x86.ActiveCfg = Release|x86
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|x86.Build.0 = Release|x86
-		{4C7B20D7-5F5C-440E-8DA3-B19A328CC8BD}.Release|x86.Deploy.0 = Release|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|ARM64.Build.0 = Debug|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x64.ActiveCfg = Debug|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x64.Build.0 = Debug|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x86.ActiveCfg = Debug|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x86.Build.0 = Debug|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|ARM64.ActiveCfg = Release|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|ARM64.Build.0 = Release|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x64.ActiveCfg = Release|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x64.Build.0 = Release|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x86.ActiveCfg = Release|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5B033CDA-6491-405C-A05B-CEE6AC4060B9}
-	EndGlobalSection
-EndGlobal
