--- WinUIGallery/WinUIGallery.sln
@@ -1,104 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.32319.34
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A98961E3-CB85-4AFC-990C-FD9F57043E7D}"
-	ProjectSection(SolutionItems) = preProject
-		..\Directory.Build.props = ..\Directory.Build.props
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WinUIGallery", "WinUIGallery.csproj", "{D78F3B56-340D-44F0-9AF1-320D473BA611}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WinUIGalleryUnitTests", "..\WinUIGalleryUnitTests\WinUIGalleryUnitTests.csproj", "{94B9B3F5-AF8E-4401-876B-115C44986161}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|ARM64 = Debug|ARM64
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Debug-Unpackaged|ARM64 = Debug-Unpackaged|ARM64
-		Debug-Unpackaged|x64 = Debug-Unpackaged|x64
-		Debug-Unpackaged|x86 = Debug-Unpackaged|x86
-		Release|ARM64 = Release|ARM64
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-		Release-Unpackaged|ARM64 = Release-Unpackaged|ARM64
-		Release-Unpackaged|x64 = Release-Unpackaged|x64
-		Release-Unpackaged|x86 = Release-Unpackaged|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|ARM64.Build.0 = Debug|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|ARM64.Deploy.0 = Debug|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x64.ActiveCfg = Debug|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x64.Build.0 = Debug|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x64.Deploy.0 = Debug|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x86.ActiveCfg = Debug|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x86.Build.0 = Debug|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug|x86.Deploy.0 = Debug|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug-Unpackaged|ARM64.ActiveCfg = Debug-Unpackaged|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug-Unpackaged|ARM64.Build.0 = Debug-Unpackaged|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug-Unpackaged|x64.ActiveCfg = Debug-Unpackaged|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug-Unpackaged|x64.Build.0 = Debug-Unpackaged|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug-Unpackaged|x86.ActiveCfg = Debug-Unpackaged|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Debug-Unpackaged|x86.Build.0 = Debug-Unpackaged|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|ARM64.ActiveCfg = Release|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|ARM64.Build.0 = Release|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|ARM64.Deploy.0 = Release|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x64.ActiveCfg = Release|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x64.Build.0 = Release|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x64.Deploy.0 = Release|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x86.ActiveCfg = Release|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x86.Build.0 = Release|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release|x86.Deploy.0 = Release|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release-Unpackaged|ARM64.ActiveCfg = Release-Unpackaged|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release-Unpackaged|ARM64.Build.0 = Release-Unpackaged|ARM64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release-Unpackaged|x64.ActiveCfg = Release-Unpackaged|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release-Unpackaged|x64.Build.0 = Release-Unpackaged|x64
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release-Unpackaged|x86.ActiveCfg = Release-Unpackaged|x86
-		{D78F3B56-340D-44F0-9AF1-320D473BA611}.Release-Unpackaged|x86.Build.0 = Release-Unpackaged|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|ARM64.Build.0 = Debug|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|ARM64.Deploy.0 = Debug|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|x64.ActiveCfg = Debug|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|x64.Build.0 = Debug|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|x64.Deploy.0 = Debug|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|x86.ActiveCfg = Debug|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|x86.Build.0 = Debug|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug|x86.Deploy.0 = Debug|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|ARM64.ActiveCfg = Debug|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|ARM64.Build.0 = Debug|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|ARM64.Deploy.0 = Debug|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|x64.ActiveCfg = Debug|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|x64.Build.0 = Debug|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|x64.Deploy.0 = Debug|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|x86.ActiveCfg = Debug|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|x86.Build.0 = Debug|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Debug-Unpackaged|x86.Deploy.0 = Debug|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|ARM64.ActiveCfg = Release|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|ARM64.Build.0 = Release|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|ARM64.Deploy.0 = Release|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|x64.ActiveCfg = Release|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|x64.Build.0 = Release|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|x64.Deploy.0 = Release|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|x86.ActiveCfg = Release|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|x86.Build.0 = Release|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release|x86.Deploy.0 = Release|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|ARM64.ActiveCfg = Release|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|ARM64.Build.0 = Release|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|ARM64.Deploy.0 = Release|ARM64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|x64.ActiveCfg = Release|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|x64.Build.0 = Release|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|x64.Deploy.0 = Release|x64
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|x86.ActiveCfg = Release|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|x86.Build.0 = Release|x86
-		{94B9B3F5-AF8E-4401-876B-115C44986161}.Release-Unpackaged|x86.Deploy.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5B033CDA-6491-405C-A05B-CEE6AC4060B9}
-	EndGlobalSection
-EndGlobal
