--- WinUIGalleryUnitTests/Properties/launchSettings.json
@@ -1,10 +1,0 @@
-{
-  "profiles": {
-    "WinUIGalleryUnitTests (Package)": {
-      "commandName": "MsixPackage"
-    },
-    "WinUIGalleryUnitTests (Unpackaged)": {
-      "commandName": "Project"
-    }
-  }
-}
