--- WinUIGalleryUnitTests/UnitTestApp.xaml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Application
-    x:Class="WinUIGalleryUnitTests.UnitTestApp"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:WinUIGalleryUnitTests">
-    <Application.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
-                <!-- Other merged dictionaries here -->
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Application.Resources>
-</Application>
