--- WinUIGalleryUnitTests/UnitTestAppWindow.xaml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Window
-    x:Class="WinUIGalleryUnitTests.UnitTestAppWindow"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:WinUIGalleryUnitTests"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Grid x:Name="rootGrid">
-
-    </Grid>
-</Window>
